Canvas: Compensate for device density
Summary:
The canvas API currently doesn't compensate for the device pixel density and therefore, the output looks largely different on different devices with different pixel densities.

Change the canvas API to compensate for device pixel density.

The additional benefit is that canvas using web `2dcontext` and the PyTorch Live canvas render the same.

Reviewed By: clarksandholtz

Differential Revision: D28668635

fbshipit-source-id: 318c5b0da97f50d3fa0cbb430875fd071662db10