Bubble up invalid canvas property exception
Summary:
The TypeScript static check will warn the user in case of an invalid canvas property for `lineJoin`, `lineCap`, and `textAlign`. However, it is still possible to set invalid values.

Previously, invalid values were ignored on the native end. This change will show a warning in `LogBox` with a message that includes valid property values.

The suggestion is to use `Promise` and reject it on an exception: https://fburl.com/sokkcogr

Reviewed By: clarksandholtz

Differential Revision: D28871773

fbshipit-source-id: 66348275b0cfc42e24088218bec1cc3fbd642431