Retain pixel density on getImageData
Summary:
Retain pixel density when calling `getImageData`. Previously, the image on canvas was drawn based on the pixel density. For example, a `ctx.rect(0, 0, 100, 100)` would draw a square in `(0,0,200,200)` if the device pixel density was `2`. Calling `getImageData(0, 0, 100, 100)` would create a bitmap of the canvas area `(0,0,200,200)` and scale it to be `(0,0,100,100)`. Re-rendering the image data would then result in a pixelated drawing.

We therefore decided to return the unscaled image data and add `naturalWidth` and `naturalHeight` props to return the true width/height `(200,200)` and `width` and height` would return `(100,100)` as if it was scaled.

This is not ideal, but an ok solution for now. However, the API is subject to change!

Reviewed By: clarksandholtz

Differential Revision: D28968720

fbshipit-source-id: 037277f3baedf1cdc20ea2596999fc31e6d321f3