Optimize textAlign
Summary:
Added local checks in JS if text align changed, and only call native setTextAlign if new value is different. Also, on the native end, it will only call the setTextAlign if the align changed compared to the current value. (setLineCap, setLineJoin, and setTextAlign are expensive native functions and should only be called when needed).

This change is similar to D29021029 for lineCap and lineJoin.

Reviewed By: clarksandholtz

Differential Revision: D29112571

fbshipit-source-id: c887f6a79752da8c863ea9beb9bf45f2be964bcb