Add example app to react-native-pytorch-core
Summary:
Adds an example app to the `react-native-pytorch-core` package. This will make it easier to develop new functions for the core package and doesn't require `yalc` or other 3p tools to pull in changes to an app outside of the core package.

The linter has warnings, but the warnings can be ignore as it is the same problem as in D28423056.

# Lint
The linter will warn about 2 gradlew.bat files. These files have third-party licenses, and we don't want to change their license.

>The suppresion syntax doesn't support REM. The line has to start with //, # or /*. You'd have to do a lot of gymnastics to get a comment line starting with that in a batch script. I think that a one-off suppression in the lint engine is overkill for this one file. I would just leave it as is.

As suggested by Andres Suarez in WP Chat.

Reviewed By: clarksandholtz

Differential Revision: D29449711

fbshipit-source-id: fee18e54ad7cd06578748b74f48db55e089a6477