Add argmax unpack for Android
Summary:
Adds `argmax` unpack to Android. The `argmax` can be used by image classification models to return the index for the most probable class.

Until PyTorch Live implements JSI, it's faster to return a single index to JavaScript rather than an array with probabilities and run argmax in JavaScript. For example, for `resnet18`, it would return an array with 1000 floating numbers

Reviewed By: clarksandholtz

Differential Revision: D29728877

fbshipit-source-id: 223922904fb2817f468706a9da1b620ffe4ca1a8