Fix rotate around a point
Summary: Rotate around a point function applied the correct rotation, but never translated the image back to it's original position.  When playing AstroBird, this was noticed because the rocket was much further right than it was supposed to be which misaligned the position of the rocket and the pipes. This diff fixes that by translating using the negative of the arguments, then rotating, then translating back using the positive value of the arguments.

Reviewed By: raedle

Differential Revision: D29700153

fbshipit-source-id: 24ce9249025576ebc31a56bc8a52be003c084a48