Use CATransformLayer to display sublayers with different transforms
Summary:
self.layer is of type CALayer which flattens all of its layers onto the same 2D plane.  We do not want this to happen on our canvas, so we must use CATransformLayer which allows sublayers to have different transforms.  Now, everything is added to drawingLayer: CATransformLayer, which in draw becomes the only sublayer of self.layer.

Each time the transformation changes, a new sublayer needs to be staged (in shapeLayers) because each layer can only have one transform.  On calls to fill or stroke, the style of stages layers are set, added to drawingLayer.  Copies of these layers are made and added to shapeLayers so that subsequent calls to fill or stroke don't override styles set by previous calls.

Reviewed By: raedle

Differential Revision: D29700154

fbshipit-source-id: 409a8903927558e817fc8d92a4fa6c6b6ee5f15c