audio packer
Summary:
Adding packer for audio for wav2vec model.

Introducing interface for using `android.media.AudioRecord` as a source of audio.

`TestAudioRecordImpl` implements this interface reading from array for tests.
Pack spec:
```
  "pack": {
    "type": "tensor_from_audio",
    "audio": "audio",
    "sample_rate": 16000,
    "duration_seconds": "$duration_seconds"
  },
```
"duration_seconds" is optional. It allows to allocate input tensor of needed size only once. If it is not specified - the assumption is that the record will fit in 10 seconds. If it will be bigger - the buffer will be resized to x2.

Reviewed By: raedle

Differential Revision: D29806963

fbshipit-source-id: bed951993769423ad54e79af6007f9210dce50a6