Optimized DrawingCanvasView
Summary: Each time a layer property was changed, an implicit `CATransaction` was created that needed to be flushed.  In order to combine these transactions, I created a `LayerData` class so calls that previously created layers (and transactions) now create layer data objects.  The draw function loops through the staged layer objects, creates layers from them, and adds them as sublayers of the root layer.  This removed the flickering and the need to call `CATransaction.flush` since draw is performed in a runloop.

Reviewed By: raedle

Differential Revision: D30170406

fbshipit-source-id: f11a7832b9a28db5dffed26ea8ee7f4cc9739d7c