Load extra files
Summary: Adding API to extract live.spec.json from model extra files. The API is similar to how the API works on PyTorch Android. The model load function returns the model implementation and the extra files param takes a mutable map/dictionary that will be filled with the model spec.

Reviewed By: hanton

Differential Revision: D30324591

fbshipit-source-id: 1ce79c35d3511b2b45a0e3493597efaa02588740