Fix CameraImageCenterCropScaleRgbNormTransform
Summary:
The camera capture broke when updating to fast unpack with center_crop/scale/image_to_tensor in a single function (CameraImageCenterCropScaleRgbNormTransform). The reasons is because the camera capture creates a jpeg instead of a YUV_420_888 image. However, the CameraImageCenterCropScaleRgbNormTransform only works with YUV_420_888.

This change adds a check if the image format is a YUV_420_888 and if it is not a YUV_420_888, it will use the default code path that performs each transformation individually (which will result in slower pack times).

Reviewed By: clarksandholtz

Differential Revision: D30861562

fbshipit-source-id: 921da2136281541cc463fa1dad9e3a74a4d4e448