Generic JSON parser for spec
Summary:
Replace fixed JSON decoder with SwiftyJSON, which allows for a more flexible decoding of the PyTorch Live spec. This change is necessary because PyTorch Live spec properties can have different data types as values for different model specifications. For example, the mean and std values for RBG normalization are arrays with three values (one for each color channel) whereas the mean and std values for grayscale normalization only has one value (because it is only one channel).

This change gives more flexibility in what is defined in the live specification format.

Reviewed By: clarksandholtz

Differential Revision: D30758684

fbshipit-source-id: c62d38856cd9153e465ed758a61dcec0b732608c