Change test regex for external fb/ import stripping
Summary:
The external OSS website build strips `fb/` imports on build, which allows us to add internal-only content. Docusaurus supports absolute paths via the `site/` prefix. However, this wasn't considered in the initial test for the `babel-plugin-transform-remove-imports`.

John Know suggested changing the regex to also catch absolute paths and strip them as well for the OSS build of the website

https://fb.workplace.com/groups/docusaurus/posts/1437442033294580?comment_id=1438213593217424&reply_comment_id=1438934826478634

Reviewed By: clarksandholtz

Differential Revision: D31083596

fbshipit-source-id: dc6971cb40dbef4519e8069fd09e56cb4d948cb8