Add unload method to MobileModelModule (#5)
Summary:
In order to change the running model to a new one, the `unload` method alow to clear all "caches" in the app to the next `preload` or `execute` will get a fresh new model from the files.
In our case, the model can be updated at runtime, so we need to be able to execute the latest one without killing the app to do so.

The iOS Parts where not tested.

EDIT (raedle): Changed function Swift function signature and `RCT_EXTERN_METHOD` definition. Tested on iOS

Pull Request resolved: https://github.com/facebookexperimental/pytorch-live/pull/5

Reviewed By: clarksandholtz

Differential Revision: D31616811

Pulled By: raedle

fbshipit-source-id: 53ef403b109a182ca882b4256f6d93a824cf1a26