Make skip-install optional
Summary:
Up until now, the iOS `pod install` was skipped. This was added because we started with Android and didn't have a working iOS version of PyTorch Live. With a working iOS version, we can remove the `--skip-install` flag. However, to allow devs unblock themselves in case of a failure, we make the `--skip-install` optional.

Example:
`npx torchlive-cli init MyProject --skip-install`

Reviewed By: liuyinglao

Differential Revision: D31723691

fbshipit-source-id: 400418272a96d18cb4ce58d26f242db690d1c5bb