use an old version of torch / torchvision for stability
Summary:
the ptl app is not compatible with the newly release pytorch (1.10) and torchvision. We pin the version for now for stability purpose, will address it after launch.

the issue can be reproduce as below

npx torchlive-cli init myfoo

cd myfoo && npx torchlive-cli run-android

watch the app can build successfully but crash when it tried to load the models.

Reviewed By: raedle

Differential Revision: D31875935

fbshipit-source-id: 6b0b1ef526ffa9e7f70358e5a8f70bf70757f6a3