check emulator version at setup
Summary:
We are installing an emulator for user to run the app on android. If the emulator is not installed, or the version is out-of-dated, the running could failed.

However, user can still run on their own emulator (with Android Studio) or on their devices.

In this diff, we give users an option to skip the installation of the latest emulator config.

Reviewed By: raedle

Differential Revision: D32090421

fbshipit-source-id: c9384f974f5849cf0259efbeb8b9c4f744c252a9