check metro occupancy when running android device
Summary:
running run-android/ios  will start metro for us in port 8081. However, if port 8081 is occupied (either by another Metro process, or by another program), it will failed.  Detect the process earlier would save user the build time of the app. Otherwise it will throw error after the

we would like to detect if the port 8081 is
1. occupied by another process, or
2, occupied by metro process running for another project

if true, we provide the user a option: whether they would like to kill the process and (re)start Metro for the current project

Y - kill the process and start metro on the app
N - exit with a message to ask user to free the port 8081

Reviewed By: raedle

Differential Revision: D32136004

fbshipit-source-id: c3bc717c2b4db7754ee2a49f5df1e719e4465a62