Allow string/url for model preload and execute
Summary:
This change adds support for passing in urls that resolve to models (which will be downloaded on the device). For example, a developer can now use models that are self-hosted such as Dropbox, GitHub, or other service that are accessible via a url.

The models are still cached locally, so if a models was downloaded once, it will still be loaded from local cache and doesn't need to be reloaded again.

The additional benefit of external urls is that models won't be part of the app bundle on release builds and reduce the initial app bundle size by large amounts depending on the model sizes used in the app.

Reviewed By: liuyinglao

Differential Revision: D32080010

fbshipit-source-id: 5a1a64ba4d8353ce04e9bcbb7de1970abf389a48