Save/load image
Summary:
This change adds API to save an `Image` to the local file system and load an `Image` from the local file system. Eventually, it will allow loading image via photo pickers (or others), run ML inference or draw them on a canvas, and save back results to the file system.

The `Playground` app saves an `Image` to the file system, logs the save path, reloads the image from that path, and then draws it on the canvas

Reviewed By: bvanderhoof

Differential Revision: D32234518

fbshipit-source-id: 89fc9eb425a01e7b3b78fac55c71ae9b3f7eb3e1