Make canvas invalidate async
Summary:
This diff changes the `Canvas.invalidate` API to be async and awaitable in JavaScript/TypeScript. It's relevant to await the canvas to finalize the drawing when handling images. For example, receiving an image from the camera, drawing it onto a canvas, invalidate (and not await), and then release the image can result in race conditions where the image is released before the canvas had a chance to draw it.

The awaitable `invalidate` function will allow the canvas to draw the image before it is released.

This was only an issue on iOS because on iOS its using `CALayer` to put together the layers to draw on the UIView. However, adding it to the `CALayer` will not yet draw the image but rather puts together the instructions for the next time the UI thread will draw the layers.

Reviewed By: clarksandholtz

Differential Revision: D32234517

fbshipit-source-id: 43f6e130d4ec21c21fc5c5708d325cd43f2c407a