fix the android sdk installer check
Summary:
Depending on how the Android ADK is installed, the sdk manager may be installed in different paths.

Currently we only check "cmdline-tools/bin/sdkmanager", which is used when the sdk is installed with "commandline-tool only"

In this diff we added another path "tools/bin/sdkmanager", which is used when the sdk is installed with android studio.

Reviewed By: raedle

Differential Revision: D32274794

fbshipit-source-id: 2a7ad039efc864217fc2a9eb1e46e5c6f8b2ef6c