Preload and unload when enter/exit demos
Summary:
Changed the `ModelPreloader` to unload models on component unmount. Then, added `ModelPreloader` to each example, which will preload models on example enter and unload on example exit.

While models are preloaded, the `ModelPreloader` will show a model screen with a loading message. This will prevent users from interacting with the example before the model finished loading, and secondly it will give a more truthful experience of the actual model runtime (otherwise the first inference or cold inference will take much longer because it will also need to wait for the model to be loaded).

Reviewed By: chrisklaiber

Differential Revision: D32329277

fbshipit-source-id: 36aeca56adafa7e8e0cce16307fa71f0cda57c8d