Fix takePicture API
Summary:
The `takePicture` API didn't call the native `receiveCommand` in the `CameraViewManager`.

The reason for this is because the React Native manager has two overloaded methods for `receiveCommand`. One has a string param for the command id and the other has an int for the command id. The latter is deprecated but currently used by the `CameraView.tsx`.

Since this method was implemented on native, it didn't trigger the `cameraView.takePicture` and therefore didn't take a picture and callback the `onCapture` callback.

Reviewed By: clarksandholtz

Differential Revision: D32420492

fbshipit-source-id: 78fbc8dfc8fee1b9315425747647a0c343a97856