add test coverage to AndroidSDK.ts
Summary:
As an example of adding unit tests to AndroidSDK.ts

According to https://stackoverflow.com/questions/45111198/how-to-mock-functions-in-the-same-module-using-jest/47976589 and
https://fb.workplace.com/groups/jesting/permalink/4898839220164654/

mocking is possible for b.func:

class a, b
a.func is calling b.func.

mocking is impossible for a.fun2:

class a
a.func1 is calling a.func2

This make it almost impossible to add tests to functions which are calling other functions of the same modules.

Another way to avoid heavy refactor may be split files so no circular or self dependancies exists

Reviewed By: raedle

Differential Revision: D32524562

fbshipit-source-id: 30421e0a1e69024aaf55682191353b9d5275a9a9