doctor command only check globally installed react-native
Summary:
A globally installed react-native could leads to unexpected behavior when developer runs pytorch-live. In our doctor command, we detect if the react-native is installed and make suggestion accordingly.

However, depending on the dir where users run the doctor command, it could report FP if react-native is installed locally (within the dir).

In this diff, we only invalidate the case where react-native is globally installed.

Reviewed By: raedle

Differential Revision: D32557819

fbshipit-source-id: 7a53b6040d61628b78f0734d1ce6e7ae899467bc