Fix Distilbert when no answer found
Summary:
The Distilbert SQuAD model will either err or return the `[CLS]` token if it cannot find an answer in the text.

An example input to provoke this issue is:

```
{"question": "How r u", "text": "Hi"}
```

This is solved by checking if the start index is smaller than zero and the start and end index are the same. In both cases, it hasn't found an answer.

Reviewed By: clarksandholtz

Differential Revision: D32583680

fbshipit-source-id: 120ab9165636386ebae5fcfbce8ecb2fbac20a8b