introduce codegen for code examples and diffs
Summary:
This codegen script keeps the examples/ directory as the source of truth, and requires a manual run to update the copies in the website/ directory.

It generates React components that can be used to display the code for an example, or the diff from that example to its previous version.

Reviewed By: raedle

Differential Revision: D32639369

fbshipit-source-id: c2429631c95f1be82bacd1af26e3f28a1f113deb