RuntimeError related to quantization solved
This commit solves a RuntimeError that occurs on ARM CPUs.

This error occurs because fbgemm backed is enabled by default as it works well
with x86 CPUs. However, ARM CPUs only work when the qnnpack backend engine is enabled.
This is the reason why in this commit qnnpack is only used when working with
ARM CPUs.

PyTorch natively supported backends:
https://pytorch.org/docs/stable/quantization.html#natively-supported-backends