init command is now independent of local python
Each time we create a project using the template, the file script.js is called.
However, this file depended on the version of Python installed locally.

In this commit, Python v3.9 is installed in MacOS via homebrew and then just this
version of Python is used.

This is crucial because we need the version 21.3.1 from PIP to install TorchVision 0.11.1.
Otherwise, we'll get that the version of TorchVision 0.11.1 does not exist.

Making the script independent of the version of Python installed locally and
installing a Python version with PIP v21.3.1 avoids having the error described above.