Delete error.log if make_models.py exits with 0
Summary:
During the template post-init phase, it runs the script that runs the `make_models.py`. This `make_models.py` will eventually export ML models and as part of the export it may download models from the PyTorch Hub. The PyTorch Hub will log the download process among other things to the stderr, which will eventually produce the error.log file. Even though the `make_models.py` script succeeds, the CLI init command will currently fail because it checks for the existance of the error.log file.

In this change, it will delete (unlink) the error.log file if the Python `make_models.py` exit with status 0 meaning it succeeded (or at least it didn't exit with a non-zero status).

Reviewed By: liuyinglao

Differential Revision: D33203002

fbshipit-source-id: 05fac7fbde27efe20a8fff5a57c9f9cb301042d2