npm publish to @latest only when release type is released
Summary:
The following case won't work --

npm publish is not triggered when a draft is saved and then published

This diff fix the case and introduce the following behavior:
1. when a new release is issued and not marked as pre-release, the npm package  with the latest tag will be updated.
2. if the new release is marked as pre-release, the npm packages with the next tag will be updated.
3. The above behavior doesn't change whether or not a draft is saved before the released.

p.s. updating the "prerelease" to "released" won't trigger the workflow, and its not desired. To make a release based on prerelease, please redo whatever steps are needed.

Reviewed By: raedle

Differential Revision: D33248441

fbshipit-source-id: f54e5fb0c2834429ed27f7c16655f6628f50944b