Change CI setting for CLI setup-dev
Summary:
The CLI build and test prematurley ends during the `setup-dev` command because it requires user input to accept license agreements.

With the recent change adding the `--yes` option and `--cocoapods-installer` option, it can not accept license agreements and sets the CocoaPods installer.

Note: the CocoaPod installer will only run if the macOS runner doesn't have CocoaPods installed already, so it may skip that step.

Reviewed By: liuyinglao

Differential Revision: D33274358

fbshipit-source-id: 099752f4f6bc2ade819fabf7c59ad84f51fefcea