Hardcoding length constraint into bert_qa model spec
Summary:
Our bert model only takes in 360 characters at a time, so this diff hardcodes that value into its spec, removes the param (`modelInputLength`) that was being passed in before, and updates the tutorials to match.

Also addresses the reported issue of the text inputs growing too large and making the UI unusable.

Reviewed By: raedle

Differential Revision: D33193948

fbshipit-source-id: 677da85c9e2b7e410ca5235e18c28d83fcfc0384