introduce `parseTensorOptions` helper method
Summary:
**This diff**
This diff introduces a `parseTensorOptions` helper method to abstract out the process of parsing tensor options from arguments. For now, we only support the `dtype` option, moving forward, this will be extended to support other keyword arguments.

This diff modifies both `torch.empty` and `torch.rand` to use this helper.

**Next diffs**
The next diffs in this stack will use this helper to add support for dtype to `torch.arange` and `torch.randint`.

Reviewed By: raedle, pranavcd

Differential Revision: D34042994

fbshipit-source-id: 1f62da1acd88a52ef6095f0e6a281e9c9f399f4f