Memory cleanup of blob data
Summary:
Retrieving the data buffer from Swift in C++ requires memcpy the data before the `torchlive::media::toBlob` leaves Swift-land; otherwise it will result in a memory BAD ACCESS. This memory needs to be cleaned up manually, which can be done in the `BlobHostObject` destructor.

This only needs to be done on iOS since Java/fbjni manages the memory itself.

Reviewed By: liuyinglao

Differential Revision: D34133606

fbshipit-source-id: 4d44a98331b9c3cc0a440bf322cb4a70c4a148c6