Fix operand parsing issue.
Summary: Previously, we assigned the address of a local variable to a pointer and then derefenced the value in a different function. This would print random garbage values after multiple runs. This diff fixes the issue by assigning the value of the parsed operand on the heap so that we persist the value across functions.

Reviewed By: liuyinglao

Differential Revision: D34288588

fbshipit-source-id: 8cfecbe7369bc868236acf8cddb119068fc04bee