add the global object torchlive and torchlive::torchvision
Summary:
The Torchvision modules provide the operators and other functionalities for image processing. In this starting diff, we implement the centerCrop function as an example.

With this diff and the following diff we will be able to call

```
let t = torch.rand([1,3,200,300]); // mock a batch of images
let centercrop = torchvision.transforms.CenterCrop(150);
let t2 = centercrop(t);
console.log(t2.shape); // [1,3,150,150]

Reviewed By: raedle

Differential Revision: D34181987

fbshipit-source-id: c6ccd5008eaee3aeb9bcc59ef0426331c7deeee0