return undefined on invalid HostObject property access
Summary:
This diff removes some exceptions thrown during property access. Generally, JavaScript objects conventionally return `undefined` for invalid property accesses.

This addresses a specific issue where a `TensorHostObject` is returned in a `Promise`. Resolving the promise was accessing the `then` property to determine if the resolved value is "then-able", which raised an exception.

While the hood was up, I also tightened up a try/catch block and installed a `std::move`.

Reviewed By: raedle

Differential Revision: D34808277

fbshipit-source-id: 708973711ab6ec7f97eb2a62e1f5bc875e1d8423