async API support
Summary:
This diff introduces support for asynchronous APIs in the JSI bindings.

JSI API installation now gets a `RuntimeExecutor`. This function can be called from any thread and will call a callback with a valid `jsi::Runtime` on a (possibly different) thread where the runtime is safe to access. Note that the callback may not be called when the runtime is no longer available.

New versions of React Native (0.66+) and TurboModules have better support for `RuntimeExecutor`, but this implementation addresses current constraints.

Renamed cpp-adapter.cpp to OnLoad.cpp consistent with fbjni convention.

While the hood was up I replaced forward declarations with torchlive.h include.

Reviewed By: raedle

Differential Revision: D34407981

fbshipit-source-id: f8013b9dbfb7b0d60985933d27b32e0bcf65483e