introduce BaseHostObject to simplify HostObject implementation
Summary:
This diff introduces a pattern to simplify defining host objects in JSI.

The new BaseHostObject provides a place to store any jsi::Value and a convenience method for storing jsi::Function values. With it, most host object subclasses will no longer need to implement HostObject::get or HostObject::getPropertyNames.

This diff applies the pattern to a function in TensorHostObject for the common case and in ModuleHostObject for an async function.

The new static functions need to be able to access the host object's internal state. TensorHostObject::tensor was already public, and I made ModuleHostObject::module_ public to match. An alternative, if we want to hide state, would be to forward-declare, e.g., class ModuleHostObjectImpl and put its implementation in the .cpp file, making the state public within that file and accessible to the static functions there.

Benefits:
* Reduce function name mentions (e.g., "abs") from 13 to 3.
* Remove a level of indentation in functions.
* Look in one place (the constructor) to see all registered functions.
* Usually no need for subclasses to define get or getPropertyNames.

Reviewed By: raedle

Differential Revision: D35153701

fbshipit-source-id: ff6d7c303cfddba71663ad4cd0ae5c50a2e25815