Tag docs for v0.1
Summary:
The PyTorch Live documentation contains the latest changes only. However, this can cause issues where the stable releases could become out of sync with the documentation.

Docusaurus supports this by versioning the state of a documentation. This change creates a version `0.1` that is in sync with the latest release `0.1.3` (and will be kept up to date with future releases of `0.1.x`).

Reviewed By: chrisklaiber

Differential Revision: D35318545

fbshipit-source-id: 1fb2139e409d510be0c18e98e0d2b079d91ccca2