use Babel plugin-transform-async-to-generator
Summary:
This fixes an issue with returning a JSI HostObject from an async function. Before, Metro used Regenerator to transform the source. Now, we use plugin-transform-async-to-generator before Regenerator runs.

This patch can be removed after metro-react-native-babel-preset is updated to 0.66.0+ which defaults to using babel/plugin-transform-async-to-generator.

The root problem is the interaction between a JSI HostObject instance, Hermes, and Regenerator. For performance, `Object.hasOwnProperty.call(aHostObject, "anything at all")` returns true in Hermes. Regenerator checks for the special property `"__await"` to determine whether to keep unwrapping a special object: https://github.com/facebook/regenerator/blob/v0.13.2/packages/regenerator-runtime/runtime.js#L160

The result is that a JSI HostObject is "unwrapped" to be returned as an undefined value. plugin-transform-async-to-generator does not have this problem and is compatible with React Native 0.64.3 / Hermes v0.7.2, the version we currently use.

Reviewed By: raedle

Differential Revision: D35443130

fbshipit-source-id: 225c42453e2d9a6943175b987aa684f3d7941f6f