convenience wrapper for async functions
Summary:
This diff makes it simpler to write functions with both synchronous and asynchronous interfaces. It splits the function into three parts: "setup" during the initial function call, "work" which is performed on a separate thread, and "resolution" which is again on React Native's JavaScript thread. The synchronous implementation simply performs all three steps on the initial call.

It also fixes bugs in the async forward function. The bug was that errors in work or in resolution were dropped and not propagated back to the Promise, which would then never be resolved.

Reviewed By: raedle

Differential Revision: D35165373

fbshipit-source-id: 2984a3b59d3592753ca2b78283b15b20aa6a8a9c