Fix image loading in Android release variant
Summary:
Images loaded with `ImageUtil.fromBundle` don't load in Android release variant builds.

The issue is that images in release variant are packaged with the apk as drawable, and therefore, need to be loaded as a resource rather than via a URL.

This change loads the image from Metro in dev mode and from drawable as a resource in release mode.

Reviewed By: clarksandholtz

Differential Revision: D35362937

fbshipit-source-id: 22b5f7d5db784ce1adbc6e89a491bb81a7a19829