pick libfbjni.so version via project jniLibs
Summary:
This diff is a stopgap until we can upgrade React Native to v0.65.

packagingOptions { pickFirst } is currently used to choose .so files from dependencies and it works differently on Mac and Linux, which chooses the version of libfbjni.so (0.0.2) packaged by RN that is not compatible with pytorch_android_lite.

This diff relies on the fact that gradle will prefer a local project's sourceSets { main { jniLibs } } } over any dependency, and include a specific copy in the local project's build.

Reviewed By: raedle

Differential Revision: D35568480

fbshipit-source-id: fe72a9b1f295c06afdb469e1ab4380a662416056