implement the tensor::item() function
Summary:
the item function is to transform a tensor of single element to a number

The pytorch JSI API doesn't support non-number tensor at this moment, so we will throw error when the dtype is not even a number. We may support non-numeric tensor in the future when we have the use case.

{F724590577}

Reviewed By: chrisklaiber

Differential Revision: D35769305

fbshipit-source-id: 7767ab3e09f732917eec20476fd49d09d7abe2b2