Retain exception message
Summary:
Capturing an exception by value will copy the exception and lower it to a generic `std::exception` losing the specific error message.

The work function in `AsyncTask` will copy the exception and thereby loose the particular exception and message.

This change retains the error message by capturing it explicitly in the lambda function capture.

Reviewed By: chrisklaiber, pranavcd

Differential Revision: D35817124

fbshipit-source-id: 9709e7602c6303f88e691722516cf5b705270468