Remove free function from BlobHostObject
Summary:
The `free` function was introduced to cleanup memory allocated to pass raw data from iOS to C++. This is not needed anymore with the `delete[]` in the `BlobHostObject` destructor.

Also, the following diff will introduce the use of smart pointer (i.e. `unique_ptr`) to manage memory for `Blob` data.

Reviewed By: chrisklaiber

Differential Revision: D35206287

fbshipit-source-id: e19bf3b4c9bbf54110445f4f4365f11635fee683