Update how we get RGB from bitmap on iOS
Summary:
**Before this change**

On iOS after getting the image either from remote url or from bundle, we store the raw bitmap data and use it to draw with CGContext and get the RGB values from it.

On Android, we don't draw it and try to get the RGB values from the bitmap directly.

**After this change**

Get rid of the discrepancy between iOS and Android, and follow how we do it on Android to get the RGB values from the bitmap directly.

Also follow this to deal with Premultiplied alpha on iOS: https://medium.com/ezefranca/dealing-with-premultiplied-alpha-on-ios-af5e1e9369bf

Reviewed By: chrisklaiber

Differential Revision: D36014185

fbshipit-source-id: 8edc6a1e3f931bcf13ff558ea829d678024f0711