Generic constraint for Module functions
Summary:
Currently, the return type for `Module.forward` and `Module.forwardSync` is `ModuleValue`, which is a union of `null | string | number | boolean | Tensor | { [key: string]: ModuleValue; } | ModuleValue[]`.

If used as a return value for a custom function, the returned function type needs to be `ModuleValue` or cast to a known union value. Otherwise, the static typing will show an error.

Use [Generic Constraints](https://www.typescriptlang.org/docs/handbook/2/generics.html#generic-constraints) for the `Module.forward` and `Module.forwardSync` functions, which will add static typing support for any type of the `ModuleValue` union type.

Also, it would allow removing the export from the `ModuleValue` type and make it an internal-only type again.

Reviewed By: chrisklaiber

Differential Revision: D36349687

fbshipit-source-id: a585331c8d7d3012e5548ff27028c957bf65659a