Propagate image generation error message to JS from native
Summary:
On `imageFromBlob()` and `imageFromTensor()`, if the width and height are not matched with the blob or tensor data, we return `null` for the generated image, and this causes an app crash downstream somewhere.
So adding the null check and throwing JS error to JS side instead of crashing the app.

Reviewed By: raedle

Differential Revision: D36533715

fbshipit-source-id: a9f1e5626e9bb489d3de2097ff43a9163af80996