Make Blob.arrayBuffer async and add size property
Summary:
This changes the `arrayBuffer()` function on the `BlobHostObject` from sync to async default to follow the [Blob](https://developer.mozilla.org/en-US/docs/Web/API/Blob) API on web.

It also adds a `size` property to return the number of bytes in the Blob similar to Blob on web.

Reviewed By: chrisklaiber

Differential Revision: D36685463

fbshipit-source-id: 2e012bc5d59da65de5e5d300955f93b0dfea5047