Properly release bridge image
Summary:
We expose `Image` via JSI `ImageHostObject` in new use cases like `fromBlob|Tensor()`.
In these cases, the underlying native image would be automatically release when JS side garbage collection is performed.

However, in order to make it compatible with bridge functions in Canvas for example, we store a copy of JSI `Image` in `JSContext`.
This copy is not released properly, this diff fixes it.

Reviewed By: raedle

Differential Revision: D36798365

fbshipit-source-id: dca8061cc62a635be453b930f8672bab83d70cf6