Image from blob
Summary:
Convert blob to `ImageHostObject`. The blob memory, size, and the image width/height are used as arguments to call from C++ to Java (MediaUtils.java) using fbjni.

The "image" is constructed in Java and under the hood is a `Bitmap`. The returned object is a `JIImage` in C++ (where `J` denotes that the object is backed by a Java object, in Java it is `IImage` interface). The returned `JIImage` is wrapped in an `Image` (implementing `IImage`), which itself will be used for the `ImageHostObject`.

Any call in JavaScript on the image will be forwarded to the `Image`, which will proxy-call the `JIImage` functions, which will call into Java (via `MediaUtils`) and finally execute the function and return its return value back to JavaScript.

Reviewed By: ansonsyfang

Differential Revision: D34873171

fbshipit-source-id: a25bc299e78e3155b2ec70ce8c1b958b86e5aeec