Select ID from NativeJSRef
Summary:
Temporary solution to make new JSI-based native media objects work with the old React Native architecture. For example, the drawImage of the canvas expects a NativeJSRef, which itself only needs to have an ID property with a UUID that resolves to an object on the native side.

The new JSI-based native media objects have this ID too to make them compatible with current approach of sending objects between native and the JS thread.

However, a JSI-based native media object might have other properties and functions that aren't serializable by the bridge.

This change adds a helper function that only selects the ID property of the ref, which guarantees that the resulting NativeJSRef object is serializable.

All of the these changes can be removed once all callsites have been migrated to use the new React Native architecture.

Reviewed By: chrisklaiber, ansonsyfang

Differential Revision: D36822315

fbshipit-source-id: f13ab952745d22ace9ad2e47e8dc5ccb09a5da7d